front_end/requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-community 
chromadb
langchain-core
google-generativeai
pdfplumber
